Split code from Version file into separate files. Version compatibility checker was added. Changed group in build.gradle to fix composite build feature in IDE.
